// types/account/accountTypes.ts
export type AccountType = {
  accountId: number; //500856714;
  accountName: string; //"817";
  accountNumber: string; //"817";
  accountNumberText: string; //"817-A";
  availableBalance: number; //0;
  backGroundColor: string; //"white";
  balance: number; //7685.04;
  image: string; //"iVBORw0KGgoAAAANSUhEUgAABM0...";
  isLoan: boolean;
  mainTextColor: string; //"#42a5f5";
  nickname: string; //"Vacation Fund";
  secondaryTextColor: string; //"#0d47a1";
  useAvailableBalance: boolean;
};